Let's also control fps. It is obvious that you want to keep the quality of input as much as possible. Besides that, we are expecting that trim etc. will work well by upsampling to high resolution.
In editing with a combination of "trim" and "concat", I now encounter a
problem where the join points do not connect smoothly. Since probably
keyframe problems are probably related to this problem, I think that
it will not solve by controlling fps alone, and I can not actually
solve it. However, as the sampling rate and the frame rate are finer,
it is no doubt that the cuttable time width will be small, so I do not
think it is meaningless.