I withdrew the remodeling done in [PR#14](https://github.com/jeorgen/align-videos-by-sound/pull/14). The biggest problem at the time of doing this was the problem of memory usage rather than processing speed.
My PC has only 4 GB in physical memory, at that time the program
immediately consumed 4 GB to handle a movie of only 5 minutes,
that made it impossible to do any other work on my PC.

PR#14 is intended to simultaneously improve memory usage and
processing speed. However, it was not a good idea to attempt
to detect by decreasing the frame rate.

"simple_compile_videos_by_sound_track" deals with, for example,
a "main" movie exceeding 20 minutes and a short movie to be inserted
in this movie.

"_find_delay" raises an exception with a high probability when
handling such a combination of real movies with
"simple_compile_videos_by_sound_track".

The main criminal of this is downsampling. The next main culprit
is max_misalignment, but it is enough to increase the default.
max_misalignment is a necessary function for movies with a time much
larger than this.

With some improvements since PR#14, the program no longer gets
messed up memory even if given a large max_misalignment,
because we made the calculation result to be cached, even if
processing is somewhat slower we can wait because it is only once.

For these reasons, I decided to stop the step of first roughly detecting.