Speaking of origin, the motivation for introducing "known_delay_ge_map" was to reduce memory usage. However, now, I would like to give "known_delay_ge_map" the role of relief manually to cases that are fooled during delay detection. When I tried quite various realistic movies, I found several patterns that inevitably find somewhat similar features. Currently this is not user controllable, so if "align" makes a mistake, the user can not do anything.
I had two problems. First, it is not very good idea to cut out media using
"-ss" in the first place. There is also a key frame problem, in short
it would be a little inaccurate if you cut it. The other one is more fatal.
Contrary to the intention of the user, it is never controlled to surely
become larger than the specified value.

As I explained at the commit time ago, thanks to a number of improvements,
this program is far gentle than before about memory usage. So there is no
longer any need to give "known_delay_ge_map" a responsibility to reduce
memory usage.

Therefore, I changed the idea of "known_delay_ge_map" fundamentally.
In a new way of thinking we never use it for cutting out with ffmpeg.
Now, we use it only for the purpose of narrowing down the answer
that "_find_delay" returns.

The original one was trying to decide only the minimum, but it should be
possible to specify the range. Therefore, I changed the name to
"known_delay_map" instead of "known_delay_ge_map".