Changed the parameter default. If these are made too large, performance problems will occur, so it is within a range that is not very radical. This change did not solve all the problems I encountered, yet "known_delay_map" does not finish the role. However, I could solve 4 out of 5 problem movies I had.
As control by "known_delay_map" increases, this is  not very useful. Especially when the chain "A is 4 minutes later than B, C is 3 minutes
later than A ..." continues, "known_delay_map" can not handle it
(because it is a simple KVS). Even at least for me it is sufficiently
large evolution that only "four problems have become one". This is because
the description of "known_delay_map" becomes simple.

Perhaps it would be possible to develop better default values. If I
find any such value, I may change it again. However, since users can
customize these, I will not rush.