Suppose there are movies A and B. It is assumed that movie A is later than either of shot start and shot end of movie B. In such a case, we could not supplement the end of movie B with the end of movie A.
"simple_compile_videos_by_sound_track" is based on the premise
that "shot" starts the shot the earliest and finishes the shot
at the latest, so it is hard to use with other patterns. This
improvement on "concat_videos_by_sound_track" is a remedy for
those patterns in "simple_compile_videos_by_sound_track".
(For example, consider movie A as shot from the MC part of the
concert but unmodering the encore, movie B omits the MC part
and shoots the encore.)