By using "aevalsrc", I made the generated script a bit more compact. This also relates to maintainability. Although it remains unchanged as to the point that the layout can not be handled except for those of 2 ch, both before and after the code change, it is still more flexible after the change.
This change seems to be a bit different in the resulting video, but I do not think this is a problem, at least because it is an indistinguishable difference for the eyes and ears of man.

Also, if this change is related to the performance of ffmpeg, we need to think carefully, but when I performed a simple measurement, the performance did not change.

Another concern was the possibility that the ffmpeg version prerequisite before and after the change may be greatly different, but since "aevalsrc" exists since ffmpeg 0.9, this will not be a problem at all.