I made relative paths available in generated scripts.
Personally I do not like absolute path descriptions, but this depends on people's preferences. There is no particular great merit in practicality. Therefore, I made it possible to control by option.

As a personal opinion, I think that relative paths are easier to maintain. Especially when you use MSYS on Windows, it is "gloomy with backslashes", it is annoying that it will become much bigger than the original name.